requirements.txt
Flask==2.3.3
pyyaml==6.0.1
python-nginx
requests==2.31.0
wtforms==3.0.1
flask-wtf==1.2.1
werkzeug==2.3.7
gunicorn==21.2.0
validators==0.22.0 